requirements.txt
pandas
biopython
torch
beautifulsoup4
transformers==4.51.3
qwen_vl_utils
accelerate
datasets
esm
requests
biotite
pymupdf4llm
pymupdf
torch
torchvision
layoutparser
SQLAlchemy==1.4.54
git+https://github.com/facebookresearch/detectron2.git@ff53992b1985b63bd3262b5a36167098e3dada02
git+https://github.com/celalp/usearch-molecules.git@cc7450978c7f448846ec527270875b71e9e20f5b
chonkie[st]
chonkie[model2vec]
pysam
lxml
pgvector
adapters
colpali-engine>=0.3.0
omnipath[graph]